feat: action type and description added, some fixes
- Now every action has it's own type. That will
help to build the advice system and calculate
productivity based on some specific actions
like sport and etc. Type is optional and
'default' by default.
- Now every action has description. It's just
a user note about action that places on the
next line. If line doesn't starts with '-' or
digit symbol, this line will be interpreted
as the description of the last action.
- validateActionString and
 calculateEfficiency functions
 moved down in the file
- Action RegExp modified to find action type
- Nature letter switch statement separated
into specific function 'natureLetterToNature'
- analyzeTime function now parses actions and
than gets the efficiency by calculateEfficiency function
- validateActionString function now has if
statement to check if action type exists in
global action types array